feat: extract message handler
Extract the message handler logic into an event handler.
