+++ src/index.ts
@@ -1,8 +1,8 @@
-import { Client, Message } from "discord.js";
+import { Client } from "discord.js";
+import { onMessage } from "./events/onMessage";
 
 const client = new Client();
 const token = process.env.TOKEN;
-const timer = 0;
 
 //connect the bot
 client
@@ -10,35 +10,4 @@ client
   .then(() => console.log("It's alive!"))
   .catch((err) => console.error(err));
 
-//message handler
-client.on("message", (message: Message) => {
-  //cooldown feature
-  if (Date.now() - timer <= 30000) {
-    return;
-  }
-
-  //bot ignores itself
-  if (message.author.id === client.user?.id) {
-    return;
-  }
-
-  //lock to #share-the-love channel
-  if (message.channel.id !== "762002255327002654") {
-    return;
-  }
-
-  const responses: string[] = [];
-  const toSay = false;
-
-  /**
-   * Push to responses
-   * make toSay true
-   * timer = date.now
-   */
-
-  // If we should say something. add the reaction and join all the responses in one message
-  if (toSay) {
-    message.react("💜");
-    message.channel.send(responses.join("\n"));
-  }
-});
+client.on("message", (message) => onMessage(message, client));
